+++ ui/src/ExportFileTemplatesList.js
@@ -44,6 +44,16 @@ class ExportFileTemplate extends Component {
     getBackEndData = async () => {
         const authorisedActivities = await getAuthorisedActivities();
         this.setState({ authorisedActivities: authorisedActivities });
+
+        this.refreshDataFromBackend(authorisedActivities);
+
+        this.interval = setInterval(
+            () => this.refreshDataFromBackend(authorisedActivities),
+            1000
+        );
+    };
+
+    refreshDataFromBackend = async (authorisedActivities) => {
         this.getExportFileTemplates(authorisedActivities);
         this.getExportFileDestinations(authorisedActivities);
     };
